requirements.txt
# requirements.txt
Flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
groq>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
openpyxl>=3.1.0
xlrd>=2.0.0